1. Working on review comments 2. Monthly expense, opening expense and advance request expense added

+++ src/main/java/com/ptsi/report/model/response/ExpenseSheetResponse.java
@@ -16,7 +16,7 @@ public class ExpenseSheetResponse {
     private LocalDate date;
     private Integer day;
     private Double openingExpense;
-    private Integer advance;
+    private Double advance;
     private Double closingExpense;
     private Double totalExpense;
     private Double totalAdvance;
